Crate dlr_wasm_interpreter

An in-place interpreter for WebAssembly 2.0

General Usage

WebAssembly (Wasm) modules must first be decoded and validated through decode_and_validate, producing a Module. This module can then be instantiated in a Store via Store::module_instantiate, creating a module instance and returning its module address, uniquely identifying this module instance within that store.

When a Store is initially created through Store::new, it is empty. This store exposes many other functions besides module instantiation to interact with it and objects allocated within it. Most notably, the Store::invoke_simple and Store::invoke methods are used to interpret Wasm code. Refer to the examples in examples/ for more information.

Example

This is an example for how to run a function exposed from a simple Wasm module (full code in examples/function_invocation.rs):

const WAT_CODE: &str = r#"
(module
    (func (export "add_one") (param $n i32) (result i32)
        local.get $n
        i32.const 1
        i32.add))
"#;

// Use the `wat` crate to convert from the text format to bytecode
let wasm_bytecode = wat::parse_str(WAT_CODE)?;

// Decode and validate the module
let module = decode_and_validate(&wasm_bytecode, &mut ())?;

// Create a new empty store
let mut store = Store::new(());

// Instantiate the module to create a module instance, returning its address
// SAFETY: There are no extern values.
let module_addr = unsafe { store.module_instantiate(&module, vec![], None) }?.module_addr;

// Get the function address of the exported add_one function
// SAFETY: The module address was returned from the same store.
let add_one_extern = unsafe { store.instance_export(module_addr, "add_one") }?;
let add_one = add_one_extern.as_func().ok_or("add_one is not a function")?;

// Invoke the function
// SAFETY: The function address was returned from the same store. There are also no address
// type parameters.
let return_values = unsafe { store.invoke_simple(add_one, vec![Value::I32(16)]) }?;
assert_eq!(*return_values, [Value::I32(17)]);

Optional Features

log: Enables logging (enables log dependency).

Modules

core 🔒

execution 🔒

instructions

All instructions, in alphanumerical order by their numeric (hex-)value

rw_spinlock

Naive implementation of spin based locking mechanisms

validation 🔒

Macros

debug

error

info

trace

unreachable_validated

warn

Structs

CustomSection

DataAddr

An address to a data instance that lives in a specific Store.

ElemAddr

An address to an element instance that lives in a specific Store.

ExternAddr

The WebAssembly specification defines an externaddr as an address to an “external” type, i.e. is a type which is managed by the embedder. For this interpreter the task of managing external objects and relating them to addresses is handed off to the user, which means that an ExternAddr can simply be seen as an integer that is opaque to Wasm code without any meaning assigned to it.

F32

F64

FuncAddr

An address to a function instance that lives in a specific Store.

FuncType

A function type

GlobalAddr

An address to a global instance that lives in a specific Store.

GlobalType

A global type

HostCall

A HostCall object contains information required for executing a specific host function.

HostResumable

A HostResumable is used to resume execution after executing its HostCall.

InstantiationOutcome

Represents a successful, possibly fueled instantiation of a module.

Limits

A limits object

MemAddr

An address to a memory instance that lives in a specific Store.

MemType

A memory type

Module

Information collected from validating a module.

ModuleAddr

An address to a module instance that lives in a specific Store.

ResultType

A result type

Store

The store represents all global state that can be manipulated by WebAssembly programs. It consists of the runtime representation of all instances of functions, tables, memories, and globals, element segments, and data segments that have been allocated during the life time of the abstract machine. https://webassembly.github.io/spec/core/exec/runtime.html#store

TableAddr

An address to a table instance that lives in a specific Store.

TableType

A table type

ValueTypeMismatchError

An error used in all TryFrom<Value> implementations for Rust types (i32, F32, Ref, …)

WasmResumable

A WasmResumable is an object used to resume execution of Wasm code.

Enums

DecodingError

Error

An opt-in error type useful for merging all error types of this crate into a single type.

ExternType

An external type

ExternVal

https://webassembly.github.io/spec/core/exec/runtime.html#external-values

NumType

A number type

Ref

RefType

A reference type

Resumable

RunState

Represents the state of a possibly interrupted resumable.

RuntimeError

TrapError

ValType

A value type

ValidationError

Value

A value at runtime. This is essentially a duplicate of ValType just with additional values.

Traits

Config

Trait that allows user specified configuration for various items during interpretation. Additionally, the types implementing this trait can act as custom user data within an interpreter instance, passed along to each method of this trait and host functions whenever they are invoked.

ValidationConfig

Functions

decode_and_validate

Type Aliases

Hostcode

Result

A definition for a Result using the optional Error type.