dlr_wasm_interpreter::core::structure

Module types

Valid Type Definitions

This module defines Rust structs for Wasm types as defined in 1, which haveh already been validated according to 2. Some structs may always be valid, while other (e.g. Limits) have invariants associated with them which are not expressed through Wasm’s type system.

Types from the Instructions Chapter

The Wasm specification defines most of its types in the types chapter1. However, block types and memargs are defined in the instructions chapter3 alongside the instructions which use them. For completeness these types are included in this Rust module as BlockType and MemArg.

WebAssembly Specification 2.0 - 2.3. Types. ↩ 1 2

WebAssembly Specification 2.0 - 3.2. Types. ↩

WebAssembly Specification 2.0 - 2.4. Instructions. ↩

Structs

FuncType

A function type

GlobalType

A global type

Limits

A limits object

MemArg

A memarg

MemType

A memory type

ResultType

A result type

TableType

A table type

VecType

A vector type

Enums

BlockType

A block type

ExternType

An external type

NumType

A number type

RefType

A reference type

ValType

A value type