dlr_wasm_interpreter

Enum DecodingError

pub enum DecodingError {

    InvalidMagic,
    InvalidBinaryFormatVersion,
    Eof,
    MalformedUtf8(Utf8Error),
    MalformedSectionTypeDiscriminator(u8),
    MalformedNumTypeDiscriminator(u8),
    MalformedVecTypeDiscriminator(u8),
    MalformedFuncTypeDiscriminator(u8),
    MalformedRefTypeDiscriminator(u8),
    MalformedValType,
    MalformedExportDescDiscriminator(u8),
    MalformedImportDescDiscriminator(u8),
    MalformedLimitsDiscriminator(u8),
    MalformedMutDiscriminator(u8),
    MalformedBlockTypeTypeIdx(i64),
    MalformedVariableLengthInteger,
    MalformedElemKindDiscriminator(u8),
    I33IsNegative,
    TooManyLocals(u64),
    SectionSizeMismatch,
    SectionOutOfOrder(SectionTy),

}

Variants

InvalidMagic

The magic number at the start of the Wasm bytecode is invalid.

InvalidBinaryFormatVersion

The binary format version at the start of the Wasm bytecode is invalid.

Eof

The end of the binary file was reached unexpectedly.

MalformedUtf8(Utf8Error)

A UTF-8 string is malformed.

MalformedSectionTypeDiscriminator(u8)

The type of a section is malformed.

MalformedNumTypeDiscriminator(u8)

The discriminator of a number type is malformed.

MalformedVecTypeDiscriminator(u8)

The discriminator of a vector type is malformed.

MalformedFuncTypeDiscriminator(u8)

The discriminator of a function type is malformed.

MalformedRefTypeDiscriminator(u8)

The discriminator of a reference type is malformed.

MalformedValType

A valtype is malformed because it is neither a number, reference nor vector type.

MalformedExportDescDiscriminator(u8)

The discriminator of an export description is malformed.

MalformedImportDescDiscriminator(u8)

The discriminator of an import description is malformed.

MalformedLimitsDiscriminator(u8)

The discriminator of a limits type is malformed.

MalformedMutDiscriminator(u8)

The discriminator of a mut type is malformed.

MalformedBlockTypeTypeIdx(i64)

Block types use a special 33-bit signed integer for encoding type indices.

MalformedVariableLengthInteger

A variable-length integer was read but it overflowed.

MalformedElemKindDiscriminator(u8)

The discriminator of an element kind is malformed.

I33IsNegative

33-bit signed integers are sometimes used to encode unsigned 32-bit integers to prevent collisions between bit patterns of different types. Therefore, 33-bit signed integers may never be negative.

TooManyLocals(u64)

A function specifies too many locals, i.e. more than 2^32 - 1

SectionSizeMismatch

A section’s contents were successfully decoded, but either too many or not enough bytes of the section’s bytecode were consumed.

SectionOutOfOrder(SectionTy)

A section with given type is out of order. All section types have a fixed order in which they must occur.