dlr_wasm_interpreter::execution

Module instructions

This module contains common definitions required by the instruction handlers, which exist in submodules of this module.

The logic for dispatching the execution of instruction handlers resides in the dispatch submodule, which itself provides multiple dispatch mechanisms. There execution is started via dispatch::run.

Additionally, the const_interpreter_loop submodule contains the execution logic for const expressions.

Modules

const_interpreter_loop

control 🔒

dispatch 🔒

The Wasm specification defines small-step reduction rules for all instructions so that a list of instructions can be reduced to some result value under some context C.

memory 🔒

numeric 🔒

parametric 🔒

reference 🔒

table 🔒

variable 🔒

vector 🔒

Structs

State 🔒

The execution state interacted with by all instructions.

Enums

InterpreterLoopOutcome

A non-error outcome of interpretation

Functions

calculate_mem_address 🔒

data_drop 🔒 ⚠

Safety

decrement_fuel 🔒

do_sidetable_control_transfer 🔒

elem_drop 🔒 ⚠

Safety

from_lanes 🔒

memory_init 🔒 ⚠

Safety

table_init 🔒 ⚠

Safety

to_lanes 🔒