dlr_wasm_interpreter::execution::instructions

Module dispatch

The Wasm specification defines small-step reduction rules for all instructions so that a list of instructions can be reduced to some result value under some context C.

An interpreter instead executes the instructions one by one with some interpreter state State. This iterative execution logic can be split into two parts: (1) instruction handlers that correspond to small-step reductions and (2) dispatching logic that decides when each instruction handler is called.

This module deals with dispatching, i.e. deciding which instruction handler to execute and when. Multiple dispatch mechanisms are available, each with their advantages and disadvantages. The dispatch mechanism to use can be configured via Config::DISPATCH_MECHANISM.

Modules

loop_call 🔒

loop_match 🔒

Macros

for_all_instructions 🔒

A macro that executes another macro with a list of all single-byte Wasm instructions, except for the extension instructions. The instructions are in an arbitrary order.

for_all_instructions_fc 🔒

A macro that executes another macro with a list of all FC Wasm instructions. The instructions are in an arbitrary order.

for_all_instructions_fd 🔒

A macro that executes another macro with a list of all FD Wasm instructions. The instructions are in an arbitrary order.

Enums

DispatchMechanism

The dispatch mechanism to use for interpreter execution.

Functions

run 🔒 ⚠

Safety