dlr_wasm_interpreter::execution::runtime_structure::memory_instances::linear_memory

Struct LinearMemory

pub struct LinearMemory<const PAGE_SIZE: usize = { crate::Limits::MEM_PAGE_SIZE as usize }> {
    pub(crate) data: Vec<u8>,
}

A linear memory is the backing data structure for a memory instance1.

It is a wrapper around a contiguous, but growable sequence of bytes. Also it provides generic methods to be build upon by memory instructions2.

TODO: Write section on why some memory instructions are implemented here vs. implemented in their instruction handlers. Also find out how we can reference the specification steps for each instruction.

WebAssembly Specification 2.0 - 4.2.9. Memory Instances. ↩

WebAssembly Specification 2.0 - 4.4.7. Memory Instructions. ↩

Fields
data: Vec<u8>
Implementations

impl<const PAGE_SIZE: usize> LinearMemory<PAGE_SIZE>

pub fn new() -> Self

Creates a new linear memory of size 0

pub fn new_with_initial_pages(pages: u16) -> Self

Creates a new zero-initialized linear memory. Its size is determined by the given number of pages and the PAGE_SIZE.

pub fn grow(&mut self, pages_to_add: u16)

Grows the current linear memory by a number of pages.

pub fn pages(&self) -> u16

Returns the size of this linear memory in pages.

pub fn len(&self) -> usize

Returns the size of this linear memory in bytes.

This size is always a multiple of PAGE_SIZE.

pub fn store<const N: usize, T: LittleEndianBytes<N>>( &mut self, index: usize, value: T, ) -> Result<(), TrapError>

Stores a T starting from the given index into this linear memory.

The T must be convertible to its little-endian byte form (as required by the LittleEndianBytes bound).

Errors

Same as LinearMemory::store_bytes

pub fn store_bytes<const N: usize>( &mut self, index: usize, bytes: [u8; N], ) -> Result<(), TrapError>

Stores a number of bytes into this linear memory starting at the given index.

Errors

TrapError::MemoryOrDataAccessOutOfBounds: The store would have been out of bounds. The memory remains unchanged.

pub fn load<const N: usize, T: LittleEndianBytes<N>>( &self, index: usize, ) -> Result<T, TrapError>

Loads a T starting from the given index into this linear memory.

The T must be convertible from its little-endian byte form (as required by the LittleEndianBytes bound).

Errors

Same as LinearMemory::load_bytes

pub fn load_bytes<const N: usize>( &self, index: usize, ) -> Result<[u8; N], TrapError>

Loads a number of bytes from this linear memory starting at the given index.

Errors

TrapError::MemoryOrDataAccessOutOfBounds: The load would have been out of bounds.

pub fn fill( &mut self, index: usize, data_byte: u8, count: usize, ) -> Result<(), TrapError>

Sets a number of bytes in this linear memory to a constant data_byte. A total of count bytes are written, starting at the given index.

Errors

TrapError::MemoryOrDataAccessOutOfBounds: The fill operation would have been out of bounds. The memory remains unchanged.

pub fn copy_within( &mut self, destination_index: usize, source_index: usize, count: usize, ) -> Result<(), RuntimeError>

Copies a number of bytes inside this linear memory from one to another location. Both the source and destination may overlap.

A total of count bytes will be copied, starting at the given source and destination indices.

Errors

TrapError::MemoryOrDataAccessOutOfBounds: The copy operation would have been out of bounds. The memory remains unchanged.

pub fn init( &mut self, destination_index: usize, source_data: &[u8], source_index: usize, count: usize, ) -> Result<(), RuntimeError>

Copies a number of bytes from the given byte slice source_data into this linear memory. A total of count bytes will be copied, starting at the given source and destination indices.

Errors

TrapError::MemoryOrDataAccessOutOfBounds: This operation would have been out of bounds. The memory remains unchanged.

pub fn data_mut(&mut self) -> &mut [u8]

Returns the data in this memory as a byte slice. The length of this slice is always a multiple of PAGE_SIZE.