dlr_wasm_interpreter::execution::runtime_structure::memory_instances::shared_linear_memory

Struct SharedLinearMemory

pub struct SharedLinearMemory {
    inner_data: RwSpinLock<Vec<AtomicU8>>,
    page_size: NonZeroUsize,
    max_pages: usize,
}

Implementation of a shared linear memory, as defined in the threads proposal1.

Unlike unshared linear memories which store their data as a vector of bytes, a shared linear memory’s backing data structure is not specified.

Instead, a number of so-called actions, which are basically the side-effects of certain instructions, are defined. The following three types of actions are implemented as of now:

rd: Reads the length or data of the memory. (Self::rd_len, Self::rd_data)

wr: Writes data into the memory. (Self::wr_data)

These actions act as a base for implementing the execution of memory instructions as described in the threads proposal1.

Implementation & Locking

This memory internally relies on a Vec<AtomicU8>. Thus, the atomic unit of information for it is a byte (u8). All byte-wise accesses to the linear memory internally occur through AtomicU8::load and AtomicU8::store, avoiding the need for an exclusive for these operations.

However, there a two reasons for why a lock must guard the inner vector of atomic bytes:

Implementing atomic stores for multi-byte values requires a global write lock. Rust’s memory model considers partially overlapping atomic operations involving a write to be undefined behavior. It is also impossible to predict whether an atomic multi-byte store operation might overlap with another operation at runtime, thereby necessitating a lock to gain temporary exclusive access.

Linear memory can grow and while shared memories must always have an upper limit, we choose not to pre-allocate the entire memory. Instead we allow growing the inner Vec allocation, which requires temporary exclusive access to move all data from the old to the new fresh allocation. The old allocation is then freed afterwards.

TODO: Does it pay of to have more fine-granular locking for multi-byte stores than a single global write lock? Would this cause issues with predictable execution times?

Unsafe Note

As the manual index checking assures all indices to be valid, there is no need to re-check. Therefore slice::get_unchecked is used access the internal AtomicU8 in the vector backing a SharedLinearMemory, implicating the use of unsafe.

To gain some confidence in the correctness of the unsafe code in this module, run miri:

cargo miri test --test memory # quick
cargo miri test # thorough

Threads Proposal Repository. ↩ 1 2

Fields
inner_data: RwSpinLock<Vec<AtomicU8>>

This vector’s length must never be 2^32 or larger.

page_size: NonZeroUsizemax_pages: usize

We have to store the maximum number of pages in this memory so that we can perform a bounds when whenever the user calls SharedLinearMemory::grow.

Implementations

impl SharedLinearMemory

pub fn new(max_pages: usize) -> Self

Create a new and empty shared linear memory. A maximum for the number of pages allocatable by this memory must be set.

pub(crate) fn new_with_page_size( max_pages: usize, page_size: NonZeroUsize, ) -> Self

pub(crate) fn new_with_initial_pages_and_page_size<T: Config>( pages: usize, max_pages: usize, page_size: NonZeroUsize, ) -> Result<Self, RuntimeError>

pub fn new_with_initial_pages<T: Config>( pages: usize, max_pages: usize, ) -> Result<Self, RuntimeError>

Create a new shared linear memory with a certain number of zero-initialized pages.

A maximum for the number of pages allocatable by this memory must also be set.

pub fn grow<T: Config>( &self, pages_to_add: usize, ) -> Result<usize, RuntimeError>

Grow the SharedLinearMemory by a number of pages, and fills the new bytes with zeros.

The entire grow operation is atomic and visible to other threads as if it was sequentially-consistent. Internally it works by temporarily acquiring exclusive access to the memory’s data storage.

pub fn page_size(&self) -> NonZeroUsize

pub fn max_pages(&self) -> usize

pub fn store<const N: usize, T: LittleEndianBytes<N>>( &self, index: usize, value: T, ord: Ordering, ) -> Result<(), RuntimeError>

At a given index, store a datum in the SharedLinearMemory

pub fn store_bytes<const N: usize>( &self, index: usize, bytes: [u8; N], ord: Ordering, ) -> Result<(), RuntimeError>

At a given index, store a datum in the SharedLinearMemory

This operation may tear.

pub fn load<const N: usize, T: LittleEndianBytes<N>>( &self, index: usize, ord: Ordering, ) -> Result<T, RuntimeError>

From a given index, load a datum from the SharedLinearMemory

pub fn load_bytes<const N: usize>( &self, index: usize, ord: Ordering, ) -> Result<[u8; N], RuntimeError>

From a given index, load a number of bytes from the SharedLinearMemory

This operation may tear.

pub fn fill( &self, index: usize, data_byte: u8, count: usize, ) -> Result<(), RuntimeError>

The memory.fill instruction. It sets a number of bytes count in this memory to a specific constant data_byte, starting at index index.

Note that the Wasm specification defines this instruction by recursion, while this implementation uses iteration.

See also

WebAssembly Specification 2.0 (with Threads Proposal) - 4.4.7 Memory Instructions - mem.fill

pub fn copy_within( &self, destination_index: usize, source_index: usize, count: usize, ) -> Result<(), RuntimeError>

The memory.copy instruction. It copies a number of bytes count from one region in this memory to another region. Starting indices must be specified for both the source and destination regions, which may also overlap.

Note that the reference specification for this implementation does not support multiple memories. Therefore, no operation to copy data between multiple memories exists.

See also

WebAssembly Specification 2.0 (with Threads Proposal) - 4.4.7 Memory Instructions - mem.copy

unsafe fn copy_byte( &self, source_index: usize, destination_index: usize, ord: Ordering, )

Copies a single byte from a given source index to some destination index.

Safety

source_index and destination_index must both be less than some length previously returned by Self::rd_len.

pub fn init( &self, destination_index: usize, source_data: &[u8], source_index: usize, count: usize, ) -> Result<(), RuntimeError>

pub fn rd_len(&self, ord: Ordering) -> usize

The rd_len action.

This action returns the length of this shared linear memory at this very specific point in time. The returned length is not the exact length of the memory, as the memory could have grown since reading of its length took place.

Note that shared linear memories can only ever grow. As a result, lengths returned by this method may only be used as a lower bound for the memory’s current size for future operations.

See also

WebAssembly Specification 2.0 (with Threads Proposal) - 4.2.17 Events - act

Self::rd_len_in_pages

pub fn rd_len_in_pages(&self, ord: Ordering) -> usize

A version of Self::rd_len that returns the length of this memory in pages instead of bytes. Refer to its documentation for more information.

pub unsafe fn rd_data( &self, index: usize, dst_byte_buffer: impl AsMut<[MaybeUninit<u8>]>, ord: Ordering, may_tear: bool, )

The rd_data action.

This action reads n bytes from index..(index + n) into a provided buffer where n is the length of that buffer in bytes.

See also

WebAssembly Specification 2.0 (with Threads Proposal) - 4.2.17 Events - act

Self::rd_data_into_array

Safety

index + n must not overflow a usize.

index + n must not be larger than some length previously returned by Self::rd_len.

n: length of destination buffer in bytes.

pub unsafe fn rd_data_into_array<const N: usize>( &self, index: usize, ord: Ordering, may_tear: bool, ) -> [u8; N]

A version of Self::rd_data that returns its data via an array.

Safety

index + N must not overflow a usize.

index + n must not be larger than some length previously returned by Self::rd_len.

unsafe fn rd_data_tearing( &self, index: usize, dst_byte_buffer: impl AsMut<[MaybeUninit<u8>]>, ord: Ordering, )

The tearing Self::rd_data variant.

Safety

index + n must not overflow a usize.

index + n must not be larger than some length previously returned by Self::rd_len.

n: length of destination buffer in bytes.

unsafe fn rd_data_no_tears( lock_guard: WriteLockGuard<'_, Vec<AtomicU8>>, index: usize, dst_byte_buffer: impl AsMut<[MaybeUninit<u8>]>, )

The non-tearing Self::rd_data variant.

This method takes a write lock guard as an argument, so that callers can potentially re-use the lock guard.

Safety

index + n must not overflow a usize.

index + n must not be larger than some length previously returned by Self::rd_len.

pub unsafe fn wr_data( &self, index: usize, bytes: impl ExactSizeIterator<Item = u8>, ord: Ordering, may_tear: bool, )

The wr action

This action writes n bytes into the memory starting at index. The bytes are read from a provided buffer in the form of an ExactSizeIterator where n is its length.

See also

WebAssembly Specification 2.0 (with Threads Proposal) - 4.2.17 Events - act

Self::rd_data_into_array

Safety

index + n must not overflow a usize.

index + n must not be larger than some length previously returned by Self::rd_len.

n: length of source buffer in bytes.

unsafe fn wr_data_tearing( &self, index: usize, bytes: impl ExactSizeIterator<Item = u8>, ord: Ordering, )

The tearing Self::wr_data variant.

Safety

index + n must not overflow a usize.

index + n must be less or equal to some length previously returned by Self::rd_len.

n: length of iterator over source buffer bytes.

unsafe fn wr_data_no_tears( lock_guard: WriteLockGuard<'_, Vec<AtomicU8>>, index: usize, bytes: impl ExactSizeIterator<Item = u8>, )

The non-tearing Self::wr_data variant.

Safety

index + n must not overflow a usize.

index + n must not be larger than some length previously returned by Self::rd_len.

n: length of iterator over source buffer bytes.

pub fn access_mut_slice<R>(&self, accessor: impl FnOnce(&mut [u8]) -> R) -> R

Allows a given closure to temporarily access the entire memory as a &mut [u8].

Note on locking

This operation exclusively locks the entire linear memory temporarily. This operation blocks until the lock is successfully acquired.