dlr_wasm_interpreter::execution::runtime_structure::store

Struct StoreInner

pub(crate) struct StoreInner {
    pub(crate) functions: AddrVec<FuncAddr, FuncInst>,
    pub(crate) tables: AddrVec<TableAddr, TableInst>,
    pub(crate) memories: AddrVec<MemAddr, MemInst>,
    pub(crate) globals: AddrVec<GlobalAddr, GlobalInst>,
    pub(crate) elements: AddrVec<ElemAddr, ElemInst>,
    pub(crate) data: AddrVec<DataAddr, DataInst>,
}

This is a Wasm store as defined by the specification. It contains all state relevant for execution.

There is a directly one-to-one relation between Store and StoreInner.

See: WebAssembly Specification 2.0 - 4.2.3 - Store

Safety

All addresses contained in a Store and its StoreInner must be valid for their associated address vectors in the same store.

Fields
functions: AddrVec<FuncAddr, FuncInst>tables: AddrVec<TableAddr, TableInst>memories: AddrVec<MemAddr, MemInst>globals: AddrVec<GlobalAddr, GlobalInst>elements: AddrVec<ElemAddr, ElemInst>data: AddrVec<DataAddr, DataInst>