dlr_wasm_interpreter

Module validation

Modules

config

error

instructions

modules

types

Decoding and Validation Logic for Types

validation_stack

Structs

Module

Information collected from validating a module.

Functions

decode_and_validate

read_all_custom_sections 🔒

Reads the next sections as long as they are custom sections and pushes them into the custom_sections vector.

validate_no_duplicate_exports 🔒