dlr_wasm_interpreter::validation::modules

Module element_segments

Functions

decode_and_validate_active_segment_offset_expr 🔒

Parse and validate an active segment offset. An active segment offset is valid if its offset expr is of type I32 and is a const expr. Additionally inserts new items to the set validation_context_refs. Validation_context_refs corresponds to C.refs in https://webassembly.github.io/spec/core/valid/conventions.html#context

decode_and_validate_generic_initializer_list 🔒

Parse and validate the initializer list of an element segment for the supplied type expected_type. An initializer list is valid with type expected_type if all of the expressions within is const and is of type expected_type. This codebase holds these as ElemItems::Exprs(RefType, Vec<Span>). Additionally inserts new items to the set validation_context_refs. validation_context_refs corresponds to C.refs in https://webassembly.github.io/spec/core/valid/conventions.html#context

decode_and_validate_shortened_initializer_list 🔒

Parse and validate a vector of func_idx’s that reflect as the initializer list of an element segment in the form of ((ref.func func_idx) end) of in the abstract syntax. An expression of such form is valid and const with type funcref if C.funcs[func_idx] exists (see link for the definition of validation context) This codebase holds these as ElemItems::RefFuncs(Vec<u32>) to sidestep the need to show Span for each expression. Additionally inserts new items to the set validation_context_refs. validation_context_refs corresponds to C.refs in https://webassembly.github.io/spec/core/valid/conventions.html#context