dlr_wasm_interpreter::validation::modules::element_segments

Function decode_and_validate_generic_initializer_list

fn decode_and_validate_generic_initializer_list(
    wasm: &mut WasmDecoder<'_>,
    expected_type: RefType,
    imported_global_types: &[GlobalType],
    c_funcs: &IdxVec<FuncIdx, TypeIdx>,
    validation_context_refs: &mut BTreeSet<FuncIdx>,
) -> Result<ElemItems, ValidationError>

Parse and validate the initializer list of an element segment for the supplied type expected_type. An initializer list is valid with type expected_type if all of the expressions within is const and is of type expected_type. This codebase holds these as ElemItems::Exprs(RefType, Vec<Span>). Additionally inserts new items to the set validation_context_refs. validation_context_refs corresponds to C.refs in https://webassembly.github.io/spec/core/valid/conventions.html#context

Returns

Ok(ElemItems::Exprs(expected_type, _)) corresponding to the parsed list if parsing & validating succeeds, Err(_) otherwise.