dlr_wasm_interpreter::execution

Module instructions

This module contains the actual interpretation loop

Modules

const_interpreter_loop

control 🔒

dispatch_tables 🔒

fc_extensions_dispatch_mod 🔒

fd_extensions_dispatch_mod 🔒

memory 🔒

numeric 🔒

parametric 🔒

reference 🔒

table 🔒

unset_mod 🔒

variable 🔒

vector 🔒

Macros

define_instruction 🔒

Structs

State 🔒

The main execution state interacted with by all instructions.

Enums

InterpreterLoopOutcome

A non-error outcome of execution of the interpreter loop

Functions

calculate_mem_address 🔒

data_drop 🔒 ⚠

Safety

decrement_fuel 🔒

do_sidetable_control_transfer 🔒

elem_drop 🔒 ⚠

Safety

fc_extensions_dispatch⚠

Safety

fd_extensions_dispatch⚠

Safety

from_lanes 🔒

memory_init 🔒 ⚠

Safety

run 🔒 ⚠

Interprets wasm native functions. Wasm parameters and Wasm return values are passed on the stack. Returns Ok(ControlFlow::Continue(())) in case execution successfully terminates, Ok(Some(required_fuel)) if execution terminates due to insufficient fuel, indicating how much fuel is required to resume with required_fuel, and [Error::RuntimeError] otherwise.

table_init 🔒 ⚠

Safety

to_lanes 🔒

unset⚠

Safety

Type Aliases

InstructionHandlerFn 🔒